wrote category schema in wrong file. Fixed mistake
